Updated training and prediction files
Updated training and prediction files to train and predict on all data. Added predict_sherlock function to mySherlock_model_helper so we are using our own prediction function. Added SplitData versions of train and predict files so data can be trained and predicted incrementally and not all at once for machines that don't have enough RAM to read in all the data at once.